+++ app/src/main/java/com/devotion/makancuy/data/model/PriceItem.kt
@@ -2,5 +2,5 @@ package com.devotion.makancuy.data.model
 
 data class PriceItem(
     val name: String,
-    val total: Int
+    val total: Double
 )
