+++ app/src/main/java/com/devotion/makancuy/presentation/detailmenu/DetailMenuViewModel.kt
@@ -24,21 +24,21 @@ class DetailMenuViewModel(
         postValue(0)
     }
 
-    val priceLiveData = MutableLiveData<Int>().apply {
-        postValue(0)
+    val priceLiveData = MutableLiveData<Double>().apply {
+        postValue(0.0)
     }
 
     fun add() {
         val count = (menuCountLiveData.value ?: 0) + 1
         menuCountLiveData.postValue(count)
-        priceLiveData.postValue(menu?.price?.times(count) ?: 0)
+        priceLiveData.postValue(menu?.price?.times(count) ?: 0.0)
     }
 
     fun minus() {
         if ((menuCountLiveData.value ?: 0) > 0) {
             val count = (menuCountLiveData.value ?: 0) - 1
             menuCountLiveData.postValue(count)
-            priceLiveData.postValue(menu?.price?.times(count) ?: 0)
+            priceLiveData.postValue(menu?.price?.times(count) ?: 0.0)
         }
     }
